+++ mingling_core/src/asset.rs
@@ -11,10 +11,13 @@ pub mod enum_tag;
 pub mod global_resource;
 
 #[doc(hidden)]
+pub mod help;
+
+#[doc(hidden)]
+pub mod lazy_resource;
+
+#[doc(hidden)]
 pub mod node;
 
 #[doc(hidden)]
 pub mod renderer;
-
-#[doc(hidden)]
-pub mod help;